+++ Presentation/CMdm.UI.Web/Controllers/MultipleRefCodeController.cs
@@ -217,7 +217,7 @@ namespace CMdm.UI.Web.Controllers
             {
                 var ids = selectedIds
                     .Split(new[] { ',' }, StringSplitOptions.RemoveEmptyEntries)
-                    .Select(x => Convert.ToInt32(x))
+                    .Select(x => Convert.ToInt64(x))
                     .ToArray();
                 docs.AddRange(_dqQueService.GetDistinctRefCodebyIds(ids));
             }
@@ -247,7 +247,7 @@ namespace CMdm.UI.Web.Controllers
             {
                 var ids = selectedIds
                     .Split(new[] { ',' }, StringSplitOptions.RemoveEmptyEntries)
-                    .Select(x => Convert.ToInt32(x))
+                    .Select(x => Convert.ToInt64(x))
                     .ToArray();
                 docs.AddRange(_dqQueService.GetMultipleRefCodebyIds(ids));
             }
