+++ Services/CMdm.Services/CustomModule/Fcmb/IActivityLogService.cs
@@ -16,13 +16,13 @@ namespace CMdm.Services.CustomModule.Fcmb
         /// </summary>
         /// <param name="recordId">recordId identifier</param>
         /// <returns>Vendor</returns>
-        IList<ActivityLog> GetActivityLogbyIds(int[] recordIds);
+        IList<ActivityLog> GetActivityLogbyIds(Int64[] recordIds);
         /// <summary>
         /// Gets a Queitem by item reference identifier
         /// </summary>
         /// <param name="recordId">que identifier</param>
         /// <returns>Vendor</returns>
-        ActivityLog GetAccLogbyId(int recordId);
+        ActivityLog GetAccLogbyId(Int64 recordId);
         /// <summary>
         /// Gets all items
         /// </summary>
