+++ Shared/CMdmi.Entities/Domain/CustomModule/Fcmb/WrongSegment.cs
@@ -10,7 +10,7 @@ namespace CMdm.Entities.Domain.CustomModule.Fcmb
     {
         [Key]
         [DatabaseGenerated(DatabaseGeneratedOption.None)]
-        public int ID { get; set; }
+        public Int64 ID { get; set; }
         public string ORGKEY { get; set; }
         public string CUST_FIRST_NAME { get; set; }
         public string CUST_MIDDLE_NAME { get; set; }
@@ -22,7 +22,7 @@ namespace CMdm.Entities.Domain.CustomModule.Fcmb
         public string SEGMENTNAME { get; set; }
         public string SUBSEGMENT { get; set; }
         public string SUBSEGMENTNAME { get; set; }
-        public int? CORP_ID { get; set; }
+        public Int64? CORP_ID { get; set; }
         public DateTime? DATE_OF_RUN { get; set; }
         public string SCHEME_CODE { get; set; }
         public string ACCOUNTNO { get; set; }
